backend/mathocr/docker_ocr/requirements.txt
fastapi
uvicorn[standard]
transformers>=4.37.0
pillow
optimum[onnxruntime]
qwen-vl-utils
python-multipart